Finished fiter functionnality and fixed pagination issue
